fixes BookingRestriction use of 0, 1, 2 as keywords
BookingRestriction used 0, 1, 2 as enums.  The caused a
conflict.  If user puts a 2 for an INT in some rule, the parser
would confuse it for a BookingRestriction and return that, giving
an unmatched rule.

This fix removes BookingRestriction enum and replaces it with
INT.  A validation rule must restrict the range of this INT (TODO).