replaces scope/linker implementation
This implementation might be easier to use in
the case of resolving references in external
resources.  Also, it is simpler to understand.

The prior implementation resolved relative task references
(!!task) by using a step counter.  This newer implementation
uses a custom ProjectQualifiedName that hides the
relativity from the internals of xtext.